* code clean: migrert vekk fra java 5 syntax
[git-p4: depot-paths = "//sktools/SKToolsKode/1.5/": change = 85723]

+++ build-utils/db-tools/src/main/java/no/statkart/sktools/utils/databasepatcher/DatabasePatcher.java
@@ -806,7 +806,7 @@ public class DatabasePatcher {
                 return false;
             }
             if (patchNumber != null ) {
-                if (patchNumber.intValue() != patchInfo.patchVersion.patchNo) {
+                if (patchNumber != patchInfo.patchVersion.patchNo) {
                     return false;
                 }
             }
